+++ submodules/IosappUI/Sources/OpenUrl.swift
@@ -232,7 +232,7 @@ private func makeInternalUrlHandler(
 }
 
 private let internetSchemes: [String] = ["http", "https"]
-private let telegramMeHosts: [String] = ["t.me", "telegram.me", "telegram.dog"]
+private let telegramMeHosts: [String] = ["asme.su", "www.asme.su"]
 
 private func handleInternetUrl(
     parsedUrl: URL,
@@ -362,7 +362,7 @@ private func appendQueryItems(to base: String, items: [URLQueryItem]) -> String 
 }
 
 private func makeIosappUrl(_ path: String, queryItems: [URLQueryItem] = []) -> String {
-    return appendQueryItems(to: "https://t.me\(path)", items: queryItems)
+    return appendQueryItems(to: "https://asme.su\(path)", items: queryItems)
 }
 
 func openExternalUrlImpl(context: AccountContext, urlContext: OpenURLContext, url: String, forceExternal: Bool, presentationData: PresentationData, navigationController: NavigationController?, dismissInput: @escaping () -> Void) {