+++ submodules/SettingsUI/Sources/Themes/ThemePreviewController.swift
@@ -514,10 +514,10 @@ public final class ThemePreviewController: ViewController {
             case .settings:
                 return
             case let .theme(theme):
-                subject = .url("https://t.me/addtheme/\(theme.slug)")
+                subject = .url("https://asme.su/addtheme/\(theme.slug)")
                 preferredAction = .default
             case let .slug(slug, _), let .themeSettings(slug, _):
-                subject = .url("https://t.me/addtheme/\(slug)")
+                subject = .url("https://asme.su/addtheme/\(slug)")
                 preferredAction = .default
             case let .media(media):
                 subject = .media(media, nil)