+++ submodules/StickerPackPreviewUI/Sources/StickerPackScreen.swift
@@ -1082,18 +1082,18 @@ private final class StickerPackContainer: ASDisplayNode {
                     links += "\n"
                 }
                 if info.id.namespace == Namespaces.ItemCollection.CloudEmojiPacks {
-                    links += "https://t.me/addemoji/\(info.shortName)"
+                    links += "https://asme.su/addemoji/\(info.shortName)"
                 } else {
-                    links += "https://t.me/addstickers/\(info.shortName)"
+                    links += "https://asme.su/addstickers/\(info.shortName)"
                 }
             }
             text = links
             shareSubject = .text(text)
         } else if let (info, _, _) = self.currentStickerPack {
             if info.id.namespace == Namespaces.ItemCollection.CloudEmojiPacks {
-                text = "https://t.me/addemoji/\(info.shortName)"
+                text = "https://asme.su/addemoji/\(info.shortName)"
             } else {
-                text = "https://t.me/addstickers/\(info.shortName)"
+                text = "https://asme.su/addstickers/\(info.shortName)"
             }
             shareSubject = .url(text)
         } else {