+++ submodules/TextFormat/Sources/GenerateTextEntities.swift
@@ -5,10 +5,8 @@ import Emoji
 
 private let whitelistedHosts: Set<String> = Set([
     "telegram.org",
-    "t.me",
-    "telegram.me",
+    "asme.su",
     "telegra.ph",
-    "telesco.pe",
     "fragment.com"
 ])